+++ src/mainloop.rs
@@ -134,7 +134,7 @@ pub fn run(config: Config, services: Vec<Service>) -> Result<()> {
                 if length.map(|length| length > MIN_LENGTH).unwrap_or(true)
                     && current_play_time > min_play_time
                 {
-                    match filter_metadata(&config, current_track) {
+                    match filter_metadata(&config, current_track, &metadata) {
                         Ok(FilterResult::Filtered(track))
                         | Ok(FilterResult::NotFiltered(track)) => {
                             for service in services.iter() {
@@ -190,7 +190,7 @@ pub fn run(config: Config, services: Vec<Service>) -> Result<()> {
                 current_track.album()
             );
 
-            match filter_metadata(&config, current_track) {
+            match filter_metadata(&config, current_track, &metadata) {
                 Ok(FilterResult::Filtered(track)) | Ok(FilterResult::NotFiltered(track)) => {
                     for service in services.iter() {
                         match service.now_playing(&track) {